+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/message/ChatEvents/DeleteChatPhoto.kt
@@ -1,5 +1,5 @@
 package dev.inmo.tgbotapi.types.message.ChatEvents
 
-import dev.inmo.tgbotapi.types.message.ChatEvents.abstracts.CommonGroupEvent
+import dev.inmo.tgbotapi.types.message.ChatEvents.abstracts.PublicChatEvent
 
-class DeleteChatPhoto: CommonGroupEvent
+class DeleteChatPhoto: PublicChatEvent