+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/message/ChatEvents/LeftChatMember.kt
@@ -1,8 +1,8 @@
 package dev.inmo.tgbotapi.types.message.ChatEvents
 
 import dev.inmo.tgbotapi.types.User
-import dev.inmo.tgbotapi.types.message.ChatEvents.abstracts.CommonGroupEvent
+import dev.inmo.tgbotapi.types.message.ChatEvents.abstracts.PublicChatEvent
 
 data class LeftChatMember(
     val user: User
-) : CommonGroupEvent
+) : PublicChatEvent