+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/message/ChatEvents/abstracts/CommonGroupEvent.kt
@@ -1,3 +1,0 @@
-package dev.inmo.tgbotapi.types.message.ChatEvents.abstracts
-
-interface CommonGroupEvent : ChannelEvent, GroupEvent, SupergroupEvent