+++ CHANGELOG.md
@@ -14,7 +14,7 @@
     * `RowBuilder` and `MatrixBuilder` now are open and available for extending
     * `MatrixBuilder#matrix` will return read-only new list instead of original internal `mutMatrix`
     * Introduced new type of events `SuccessfulPaymentEvent` instead of putting of payment inside of message
-    * New type of events union: `CommonGroupEvent`
+    * New type of events union: `PublicChatEvent`
 * `Utils`:
     * Two new dsl:
         * `inlineKeyboard` for creating `InlineKeyboardMarkup`